+++ src/components/Accordion/Accordion.tsx
@@ -8,6 +8,7 @@ export interface AccordionPanelProps {
   id: string
   details: React.ReactNode
   summary: React.ReactNode
+  addDomId?: boolean
 }
 
 export interface AccordionPanelInternalProps extends Omit<AccordionProps, 'children' | 'activeId' | 'className'> {
@@ -15,11 +16,21 @@ export interface AccordionPanelInternalProps extends Omit<AccordionProps, 'child
   togglePanel(): void
 }
 
-function AccordionPanel(props: AccordionPanelProps & AccordionPanelInternalProps): React.ReactElement {
-  const { summary, details, togglePanel, isOpen, id, collapseProps } = props
+export const AccordionPanelWithRef = React.forwardRef(AccordionPanel)
+
+function AccordionPanel(
+  props: AccordionPanelProps & AccordionPanelInternalProps,
+  ref: React.Ref<HTMLDivElement>
+): React.ReactElement {
+  const { summary, details, togglePanel, isOpen, id, collapseProps, addDomId } = props
 
   return (
-    <div data-testid={`${id}-panel`} className={cx(css.panel, props.panelClassName)} data-open={isOpen}>
+    <div
+      ref={ref}
+      data-testid={`${id}-panel`}
+      className={cx(css.panel, props.panelClassName)}
+      data-open={isOpen}
+      id={addDomId ? `${id}-panel` : undefined}>
       <div data-testid={`${id}-summary`} onClick={togglePanel} className={cx(css.summary, props.summaryClassName)}>
         <div className={cx({ [css.label]: typeof summary === 'string' })}>{summary}</div>
         <div className={css.chevron} />
@@ -45,20 +56,14 @@ export interface AccordionProps {
 }
 
 export function Accordion(props: AccordionProps): React.ReactElement {
-  const { children, allowMultiOpen, className } = props
+  const { children, allowMultiOpen, className, activeId, ...rest } = props
   const [activePanels, setActivePanels] = React.useState<Record<string, boolean>>(
-    typeof props.activeId === 'string' ? { [props.activeId]: true } : {}
+    typeof activeId === 'string' ? { [activeId]: true } : {}
   )
 
   function togglePanels(id: string) {
     return () => {
-      setActivePanels(prev => {
-        if (prev[id]) {
-          return { ...(allowMultiOpen ? prev : {}), [id]: false }
-        } else {
-          return { ...(allowMultiOpen ? prev : {}), [id]: true }
-        }
-      })
+      setActivePanels(prev => ({ ...(allowMultiOpen ? prev : {}), [id]: !prev[id] }))
     }
   }
 
@@ -71,8 +76,9 @@ export function Accordion(props: AccordionProps): React.ReactElement {
           const { props: childProps } = child as React.ReactElement<AccordionPanelProps>
 
           return (
-            <AccordionPanel
+            <AccordionPanelWithRef
               key={childProps.id}
+              {...rest}
               {...childProps}
               isOpen={!!activePanels[childProps.id]}
               togglePanel={togglePanels(childProps.id)}