+++ src/components/MultiTypeInput/MultiTypeInput.stories.tsx
@@ -22,3 +22,45 @@ export const TextInput: Story<MultiTextInputProps> = args => <MultiTextInput {..
 export const SelectInput: Story<MultiTypeInputProps> = args => <MultiTypeInput {...args} />
 
 export const MultiSelectInput: Story<MultiSelectTypeInputProps> = args => <MultiSelectTypeInput {...args} />
+
+TextInput.args = {
+  mentionsInfo: {
+    data: (done: (arg0: string[]) => void) =>
+      done([
+        'app.name',
+        'app.description',
+        'pipeline.name',
+        'pipeline.description',
+        'pipeline.identifier',
+        'pipeline.stage.qa.displayName'
+      ])
+  }
+}
+
+SelectInput.args = {
+  mentionsInfo: {
+    data: (done: (arg0: string[]) => void) =>
+      done([
+        'app.name',
+        'app.description',
+        'pipeline.name',
+        'pipeline.description',
+        'pipeline.identifier',
+        'pipeline.stage.qa.displayName'
+      ])
+  }
+}
+
+MultiSelectInput.args = {
+  mentionsInfo: {
+    data: (done: (arg0: string[]) => void) =>
+      done([
+        'app.name',
+        'app.description',
+        'pipeline.name',
+        'pipeline.description',
+        'pipeline.identifier',
+        'pipeline.stage.qa.displayName'
+      ])
+  }
+}